pyproject.toml
[project]
name = "working-mais"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.3.1",
    "matplotlib>=3.10.8",
    "networkx>=3.6.1",
    "numexpr>=2.14.1",
    "numpy-indexed>=0.3.7",
    "pandas>=2.3.3",
    "scikit-learn>=1.8.0",
    "tqdm>=4.67.1",
]
